requirements.txt
langgraph
langchain
google-genai
fastapi
uvicorn
python-dotenv
streamlit
requests
uuid
pyttsx3